Abstract ERC20 transactions from CoreWrapper
